Split eval.py into modules
