+++ cspell.json
@@ -7,8 +7,11 @@
         "AOAI",
         "azuretools",
         "dataframe",
+        "groundtruth",
+        "jaccard",
         "keyvault",
-        "spacy"
+        "spacy",
+        "textdistance"
     ],
     "ignoreWords": [
         "BLAS",
