+++ rag_experiment_accelerator/config/path_config.py
@@ -4,6 +4,16 @@ from rag_experiment_accelerator.config import paths
 from rag_experiment_accelerator.config.base_config import BaseConfig
 
 
+class Paths:
+    ARTIFACTS_DIR = "artifacts"
+    DATA_DIR = "data"
+    EVAL_DATA_FILE = "eval_data.jsonl"
+    GENERATED_INDEX_NAMES_FILE = "generated_index_names.jsonl"
+    QUERY_DATA_DIR = "query_data"
+    EVAL_DATA_DIR = "eval_score"
+    SAMPLING_OUTPUT_DIR = "sampling"
+
+
 @dataclass
 class PathConfig(BaseConfig):
     artifacts_dir: str = ""
@@ -18,32 +28,34 @@ class PathConfig(BaseConfig):
         self._config_dir = os.path.dirname(config_file_path)
 
         if not self.artifacts_dir:
-            self.artifacts_dir = os.path.join(self._config_dir, "artifacts")
+            self.artifacts_dir = os.path.join(self._config_dir, Paths.ARTIFACTS_DIR)
         paths.try_create_directory(self.artifacts_dir)
 
         if data_dir:
             self.data_dir = data_dir
         elif not self.data_dir:
-            self.data_dir = os.path.join(self._config_dir, "data")
+            self.data_dir = os.path.join(self._config_dir, Paths.DATA_DIR)
 
         if not self.eval_data_file:
-            self.eval_data_file = os.path.join(self.artifacts_dir, "eval_data.jsonl")
+            self.eval_data_file = os.path.join(self.artifacts_dir, Paths.EVAL_DATA_FILE)
 
         if not self.generated_index_names_file:
             self.generated_index_names_file = os.path.join(
-                self.artifacts_dir, "generated_index_names.jsonl"
+                self.artifacts_dir, Paths.GENERATED_INDEX_NAMES_FILE
             )
 
         if not self.query_data_dir:
-            self.query_data_dir = os.path.join(self.artifacts_dir, "query_data")
+            self.query_data_dir = os.path.join(self.artifacts_dir, Paths.QUERY_DATA_DIR)
         paths.try_create_directory(self.query_data_dir)
 
         if not self.eval_data_dir:
-            self.eval_data_dir = os.path.join(self.artifacts_dir, "eval_score")
+            self.eval_data_dir = os.path.join(self.artifacts_dir, Paths.EVAL_DATA_DIR)
         paths.try_create_directory(self.eval_data_dir)
 
         if not self.sampling_output_dir:
-            self.sampling_output_dir = os.path.join(self.artifacts_dir, "sampling")
+            self.sampling_output_dir = os.path.join(
+                self.artifacts_dir, Paths.SAMPLING_OUTPUT_DIR
+            )
         paths.try_create_directory(self.sampling_output_dir)
 
     def sampled_cluster_predictions_path(self, optimum_k: int) -> str:
