+++ rag_experiment_accelerator/evaluation/eval.py
@@ -1,32 +1,36 @@
-import ast
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from contextlib import ExitStack
 import os
 import warnings
 
-import evaluate
 import mlflow
 import pandas as pd
-import plotly.express as px
-import plotly.graph_objects as go
-import plotly.subplots as sp
-import textdistance
+
 from dotenv import load_dotenv
-from fuzzywuzzy import fuzz
 from numpy import mean
-from sentence_transformers import SentenceTransformer
-from sklearn.metrics.pairwise import cosine_similarity
 
 from rag_experiment_accelerator.artifact.handlers.query_output_handler import (
     QueryOutputHandler,
 )
 from rag_experiment_accelerator.config.config import Config
 from rag_experiment_accelerator.config.index_config import IndexConfig
-from rag_experiment_accelerator.llm.prompt import (
-    llm_answer_relevance_instruction,
-    llm_context_recall_instruction,
-    llm_context_precision_instruction,
+from rag_experiment_accelerator.evaluation import plain_metrics
+from rag_experiment_accelerator.evaluation.llm_based_metrics import (
+    compute_llm_based_score,
+    lower_and_strip,
 )
+from rag_experiment_accelerator.evaluation.plot_metrics import (
+    draw_hist_df,
+    draw_search_chart,
+    generate_metrics,
+    plot_apk_scores,
+    plot_map_scores,
+    plot_mapk_scores,
+)
+from rag_experiment_accelerator.evaluation.transformer_based_metrics import (
+    compute_transformer_based_score,
+)
+
 from rag_experiment_accelerator.llm.response_generator import ResponseGenerator
 from rag_experiment_accelerator.utils.logging import get_logger
 from rag_experiment_accelerator.config.environment import Environment
@@ -36,489 +40,19 @@ logger = get_logger(__name__)
 load_dotenv()
 warnings.filterwarnings("ignore")
 
-algs = textdistance.algorithms
-
-
-def lower(text):
-    """
-    Converts the input text to lowercase.
-
-    Args:
-        text (str): The text to convert to lowercase.
-
-    Returns:
-        str: The input text in lowercase.
-    """
-    return text.lower()
-
-
-def remove_spaces(text):
-    """
-    Removes leading and trailing spaces from a string.
-
-    Args:
-        text (str): The string to remove spaces from.
-
-    Returns:
-        str: The input string with leading and trailing spaces removed.
-    """
-    return text.strip()
-
-
-def lower_and_strip(text):
-    """
-    Converts the input to lowercase without spaces or empty string if None.
-
-    Args:
-        text (str): The string to format.
-
-    Returns:
-        str: The formatted input string.
-    """
-    if text is None:
-        return ""
-    else:
-        return text.lower().strip()
-
-
-# https://huggingface.co/spaces/evaluate-metric/bleu
-def bleu(predictions, references):
-    bleu = evaluate.load("bleu")
-
-    results = bleu.compute(predictions=predictions, references=references, max_order=2)
-    # multiplying by 100 to maintain consistency with previous implementation
-    return results["bleu"] * 100
-
-
-def fuzzy(doc1, doc2):
-    """
-    Calculates the fuzzy score between two documents.
-
-    Parameters:
-        doc1 (str): The first document to compare.
-        doc2 (str): The second document to compare.
-
-    Returns:
-        int: The fuzzy score between the two documents.
-    """
-    differences = fuzzy_compare_values(doc1, doc2)
-
-    return int(sum(differences) / len(differences))
-
-
-def fuzzy_compare_values(value1, value2) -> list[float]:
-    """
-    Compares two values using fuzzy string matching and appends the similarity score to a list of differences.
-
-    Args:
-        value1 (str): The first value to compare.
-        value2 (str): The second value to compare.
-
-    Returns:
-        A list of the similarity scores.
-    """
-    similarity_score = [fuzz.token_set_ratio(value1, value2)]
-    return similarity_score
-
-
-def compare_semantic_document_values(doc1, doc2, model_type):
-    """
-    Compares the semantic values of two documents and returns the percentage of differences.
-
-    Args:
-        doc1 (str): The first document to compare.
-        doc2 (str): The second document to compare.
-        model_type (SentenceTransformer): The SentenceTransformer model to use for comparison.
-
-    Returns:
-        int: The percentage of differences between the two documents.
-    """
-    differences = semantic_compare_values(doc1, doc2, model_type)
-
-    return int(sum(differences) / len(differences))
-
-
-def semantic_compare_values(
-    value1: str,
-    value2: str,
-    model_type: SentenceTransformer,
-) -> list[float]:
-    """
-    Computes the semantic similarity between two values using a pre-trained SentenceTransformer model.
-
-    Args:
-        value1 (str): The first value to compare.
-        value2 (str): The second value to compare.
-        model_type (SentenceTransformer): The pre-trained SentenceTransformer model to use for encoding the values.
-
-    Returns:
-        A list of the similarity scores.
-    """
-    embedding1 = model_type.encode([str(value1)])
-    embedding2 = model_type.encode([str(value2)])
-    similarity_score = cosine_similarity(embedding1, embedding2)
-
-    return [similarity_score * 100]
-
-
-def levenshtein(value1, value2):
-    """
-    Calculates the Levenshtein distance between two strings and returns the normalized similarity score as a percentage.
-
-    Args:
-        value1 (str): The first string to compare.
-        value2 (str): The second string to compare.
-
-    Returns:
-        int: The normalized similarity score as a percentage.
-    """
-    score = int(algs.levenshtein.normalized_similarity(value1, value2) * 100)
-    return score
-
-
-def jaccard(value1, value2):
-    """
-    Calculates the Jaccard similarity score between two sets of values.
-
-    Args:
-        value1 (set): The first set of values.
-        value2 (set): The second set of values.
-
-    Returns:
-        int: The Jaccard similarity score between the two sets of values, as a percentage.
-    """
-    score = int(algs.jaccard.normalized_similarity(value1, value2) * 100)
-    return score
-
-
-def hamming(value1, value2):
-    """
-    Calculates the Hamming similarity score between two values.
-
-    Args:
-        value1 (str): The first value to compare.
-        value2 (str): The second value to compare.
-
-    Returns:
-        int: The Hamming similarity score between the two values, as a percentage.
-    """
-    score = int(algs.hamming.normalized_similarity(value1, value2) * 100)
-    return score
-
-
-def jaro_winkler(value1, value2):
-    """
-    Calculates the Jaro-Winkler similarity score between two strings.
-
-    Args:
-        value1 (str): The first string to compare.
-        value2 (str): The second string to compare.
-
-    Returns:
-        int: The Jaro-Winkler similarity score between the two strings, as an integer between 0 and 100.
-    """
-    score = int(algs.jaro_winkler.normalized_similarity(value1, value2) * 100)
-    return score
-
-
-def cosine(value1, value2):
-    """
-    Calculates the cosine similarity (Ochiai coefficient) between two strings
-    using token-frequency vectors
-
-    https://en.wikipedia.org/wiki/Cosine_similarity.
-
-    Args:
-        value1 (list): The first vector.
-        value2 (list): The second vector.
-
-    Returns:
-        int: The cosine similarity score between the two vectors, as a percentage.
-    """
-    score = int(algs.cosine.normalized_similarity(value1, value2) * 100)
-    return score
-
-
-def lcsseq(value1, value2):
-    """
-    Computes the longest common subsequence (LCS) similarity score between two input strings.
-
-    Args:
-        value1 (str): The first input string.
-        value2 (str): The second input string.
-
-    Returns:
-        int: The LCS similarity score between the two input strings, as a percentage (0-100).
-    """
-    score = int(algs.lcsseq.normalized_similarity(value1, value2) * 100)
-    return score
-
-
-def lcsstr(value1, value2):
-    """
-    Calculates the longest common substring (LCS) similarity score between two strings.
-
-    Args:
-        value1 (str): The first string to compare.
-        value2 (str): The second string to compare.
-
-    Returns:
-        int: The LCS similarity score between the two strings, as a percentage (0-100).
-    """
-    score = int(algs.lcsstr.normalized_similarity(value1, value2) * 100)
-    return score
-
-
-def llm_answer_relevance(
-    response_generator: ResponseGenerator, question, answer
-) -> float:
-    """
-    Scores the relevancy of the answer according to the given question.
-    Answers with incomplete, redundant or unnecessary information is penalized.
-    Score can range from 0 to 1 with 1 being the best.
-
-    Args:
-        question (str): The question being asked.
-        answer (str): The generated answer.
-
-    Returns:
-        double: The relevancy score generated between the question and answer.
-
-    """
-    result = response_generator.generate_response(
-        llm_answer_relevance_instruction, text=answer
-    )
-    if result is None:
-        logger.warning("Unable to generate answer relevance score")
-        return 0.0
-
-    model = SentenceTransformer("sentence-transformers/all-MiniLM-L6-v2")
-
-    embedding1 = model.encode([str(question)])
-    embedding2 = model.encode([str(result)])
-    similarity_score = cosine_similarity(embedding1, embedding2)
-
-    return float(similarity_score[0][0] * 100)
-
-
-def llm_context_precision(
-    response_generator: ResponseGenerator, question, retrieved_contexts
-) -> float:
-    """
-    Computes precision by assessing whether each retrieved context is useful for answering a question.
-    Only considers the presence of relevant chunks in the retrieved contexts, but doesn't take into
-    account their ranking order.
-
-    Args:
-        question (str): The question being asked.
-        retrieved_contexts (list[str]): The list of retrieved contexts for the query.
-
-    Returns:
-        double: proportion of relevant chunks retrieved for the question
-    """
-    relevancy_scores = []
-
-    for context in retrieved_contexts:
-        result: str | None = response_generator.generate_response(
-            llm_context_precision_instruction,
-            context=context,
-            question=question,
-        )
-        llm_judge_response = lower_and_strip(result)
-        # Since we're only asking for one response, the result is always a boolean 1 or 0
-        if llm_judge_response == "yes":
-            relevancy_scores.append(1)
-        elif llm_judge_response == "no":
-            relevancy_scores.append(0)
-        else:
-            logger.warning("Unable to generate context precision score")
-
-    logger.debug(relevancy_scores)
-
-    if not relevancy_scores:
-        logger.warning("Unable to compute average context precision")
-        return -1
-    else:
-        return (sum(relevancy_scores) / len(relevancy_scores)) * 100
-
-
-def llm_context_recall(
-    response_generator: ResponseGenerator,
-    question,
-    groundtruth_answer,
-    retrieved_contexts,
-):
-    """
-    Estimates context recall by estimating TP and FN using annotated answer (ground truth) and retrieved context.
-    Context_recall values range between 0 and 1, with higher values indicating better performance.
-    To estimate context recall from the ground truth answer, each sentence in the ground truth answer is analyzed to determine
-    whether it can be attributed to the retrieved context or not. In an ideal scenario, all sentences in the ground truth answer
-    should be attributable to the retrieved context. The formula for calculating context recall is as follows:
-    context_recall = GT sentences that can be attributed to context / nr sentences in GT
-
-    Code adapted from https://github.com/explodinggradients/ragas
-    Copyright [2023] [Exploding Gradients]
-    under the Apache License (see evaluation folder)
-
-    Args:
-        question (str): The question being asked
-        groundtruth_answer (str): The ground truth ("output_prompt")
-        retrieved_contexts (list[str]): The list of retrieved contexts for the query
-
-    Returns:
-        double: The context recall score generated between the ground truth (expected) and context.
-    """
-    context = "\n".join(retrieved_contexts)
-    prompt = (
-        "\nquestion: "
-        + question
-        + "\ncontext: "
-        + context
-        + "\nanswer: "
-        + groundtruth_answer
-    )
-    result = response_generator.generate_response(
-        sys_message=llm_context_recall_instruction,
-        prompt=prompt,
-    )
-    good_response = '"Attributed": "1"'
-    bad_response = '"Attributed": "0"'
-
-    return (
-        result.count(good_response)
-        / (result.count(good_response) + result.count(bad_response))
-    ) * 100
-
-
-def generate_metrics(experiment_name, run_id, mlflow_client):
-    """
-    Generates metrics for a given experiment and run ID.
-
-    Args:
-        experiment_name (str): The name of the experiment.
-        run_id (int): The ID of the run.
-        mlflow_client (mlflow.MlflowClient): The MLflow client to use for logging the metrics.
-
-    Returns:
-        None
-    """
-    experiment = dict(mlflow_client.get_experiment_by_name(experiment_name))
-    runs_list = mlflow_client.search_runs([experiment["experiment_id"]])
-
-    models_metrics = {}
-    metrics_to_plot = []
-    runs_id_to_plot = []
-
-    if len(runs_list) > 0:
-        for run in runs_list:
-            run_dict = run.to_dictionary()
-            single_run_id = run_dict["info"]["run_id"]
-            runs_id_to_plot.append(single_run_id)
-            if run.data.params.get("run_metrics", {}) != {}:
-                metrics = ast.literal_eval(run.data.params["run_metrics"])
-
-                for metric_type, metric_value in metrics.items():
-                    if models_metrics.get(metric_type, {}) == {}:
-                        metrics_to_plot.append(metric_type)
-                        models_metrics[metric_type] = {}
-
-                    models_metrics[metric_type][single_run_id] = metric_value
-                logger.debug(models_metrics)
-    else:
-        current_run = mlflow_client.get_run(run_id)
-        if current_run.data.params.get("run_metrics", {}) != {}:
-            metrics = ast.literal_eval(current_run.data.params["run_metrics"])
-            for metric_type, metric_value in metrics.items():
-                if models_metrics.get(metric_type, {}) == {}:
-                    metrics_to_plot.append(metric_type)
-                    models_metrics[metric_type] = {}
-
-                models_metrics[metric_type][run_id] = metric_value
-
-    x_axis = []
-    y_axis = []
-
-    fig = go.Figure()
-
-    for metric in metrics_to_plot:
-        for key, value in models_metrics[metric].items():
-            x_axis.append(key)
-            y_axis.append(value)
-
-        label = key
-        px.line(x_axis, y_axis)
-        fig.add_trace(go.Scatter(x=x_axis, y=y_axis, mode="lines+markers", name=label))
-
-        fig.update_layout(
-            xaxis_title="run name", yaxis_title=metric, font=dict(size=15)
-        )
-
-        plot_name = metric + ".html"
-        mlflow_client.log_figure(run_id, fig, plot_name)
-
-        fig.data = []
-        fig.layout = {}
-        x_axis = []
-        y_axis = []
-
-
-def draw_hist_df(df, run_id, mlflow_client):
-    """
-    Draw a histogram of the given dataframe and log it to the specified run ID.
-
-    Args:
-        df (pandas.DataFrame): The dataframe to draw the histogram from.
-        run_id (str): The ID of the run to log the histogram to.
-        mlflow_client (mlflow.MlflowClient): The MLflow client to use for logging the histogram.
-
-    Returns:
-        None
-    """
-    fig = px.bar(
-        x=df.columns,
-        y=df.values.tolist(),
-        title="metric comparison",
-        color=df.columns,
-        labels=dict(x="Metric Type", y="Score", color="Metric Type"),
-    )
-    plot_name = "all_metrics_current_run.html"
-    mlflow_client.log_figure(run_id, fig, plot_name)
-
-
-def plot_apk_scores(df, run_id, mlflow_client):
-    fig = px.line(df, x="k", y="score", title="AP@k scores", color="search_type")
-    plot_name = "average_precision_at_k.html"
-    mlflow_client.log_figure(run_id, fig, plot_name)
-
-
-# maybe pull these 2 above and below functions into a single one
-def plot_mapk_scores(df, run_id, mlflow_client):
-    fig = px.line(df, x="k", y="map_at_k", title="MAP@k scores", color="search_type")
-    plot_name = "mean_average_precision_at_k.html"
-    mlflow_client.log_figure(run_id, fig, plot_name)
-
-
-def plot_map_scores(df, run_id, mlflow_client):
-    fig = px.bar(df, x="search_type", y="mean", title="MAP scores", color="search_type")
-    plot_name = "mean_average_precision_scores.html"
-    mlflow_client.log_figure(run_id, fig, plot_name)
-
 
 def compute_metrics(
-    response_generator: ResponseGenerator,
+    metric_type,
     question,
     actual,
     expected,
+    response_generator: ResponseGenerator,
     retrieved_contexts,
-    metric_type,
 ):
     """
     Computes a score for the similarity between two strings using a specified metric.
 
     Args:
-        actual (str): The first string to compare.
-        expected (str): The second string to compare.
-        retrieved_contexts (list[str]): The list of retrieved contexts for the query.
         metric_type (str): The type of metric to use for comparison. Valid options are:
             - "lcsstr": Longest common substring
             - "lcsseq": Longest common subsequence
@@ -537,79 +71,35 @@ def compute_metrics(
             - "llm_context_precision": Verifies whether or not a given context is useful for answering a question.
             - "llm_answer_relevance": Scores the relevancy of the answer according to the given question.
             - "llm_context_recall": Scores context recall by estimating TP and FN using annotated answer (ground truth) and retrieved context.
+        question (str): question text
+        actual (str): The first string to compare.
+        expected (str): The second string to compare.
+        response_generator (ResponseGenerator): The response generator to use for generating responses.
+        retrieved_contexts (list[str]): The list of retrieved contexts for the query.
+
 
     Returns:
         float: The similarity score between the two strings, as determined by the specified metric.
     """
-    match metric_type:
-        case "lcsstr":
-            score = lcsstr(actual, expected)
-        case "lcsseq":
-            score = lcsseq(actual, expected)
-        case "cosine":
-            score = cosine(actual, expected)
-        case "jaro_winkler":
-            score = jaro_winkler(actual, expected)
-        case "hamming":
-            score = hamming(actual, expected)
-        case "jaccard":
-            score = jaccard(actual, expected)
-        case "levenshtein":
-            score = levenshtein(actual, expected)
-        case "fuzzy":
-            score = fuzzy(actual, expected)
-        case "bert_all_MiniLM_L6_v2":
-            all_MiniLM_L6_v2 = SentenceTransformer(
-                "sentence-transformers/all-MiniLM-L6-v2"
+
+    plain_metric_func = getattr(plain_metrics, metric_type, None)
+    if plain_metric_func:
+        return plain_metric_func(actual, expected)
+
+    try:
+        score = compute_transformer_based_score(actual, expected, metric_type)
+    except KeyError:
+        try:
+            score = compute_llm_based_score(
+                metric_type,
+                question,
+                actual,
+                expected,
+                response_generator,
+                retrieved_contexts,
             )
-            score = compare_semantic_document_values(actual, expected, all_MiniLM_L6_v2)
-        case "bert_base_nli_mean_tokens":
-            base_nli_mean_tokens = SentenceTransformer(
-                "sentence-transformers/bert-base-nli-mean-tokens"
-            )
-            score = compare_semantic_document_values(
-                actual, expected, base_nli_mean_tokens
-            )
-        case "bert_large_nli_mean_tokens":
-            large_nli_mean_tokens = SentenceTransformer(
-                "sentence-transformers/bert-large-nli-mean-tokens"
-            )
-            score = compare_semantic_document_values(
-                actual, expected, large_nli_mean_tokens
-            )
-        case "bert_large_nli_stsb_mean_tokens":
-            large_nli_stsb_mean_tokens = SentenceTransformer(
-                "sentence-transformers/bert-large-nli-stsb-mean-tokens"
-            )
-            score = compare_semantic_document_values(
-                actual, expected, large_nli_stsb_mean_tokens
-            )
-        case "bert_distilbert_base_nli_stsb_mean_tokens":
-            distilbert_base_nli_stsb_mean_tokens = SentenceTransformer(
-                "sentence-transformers/distilbert-base-nli-stsb-mean-tokens"
-            )
-            score = compare_semantic_document_values(
-                actual, expected, distilbert_base_nli_stsb_mean_tokens
-            )
-        case "bert_paraphrase_multilingual_MiniLM_L12_v2":
-            paraphrase_multilingual_MiniLM_L12_v2 = SentenceTransformer(
-                "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2"
-            )
-            score = compare_semantic_document_values(
-                actual, expected, paraphrase_multilingual_MiniLM_L12_v2
-            )
-        case "llm_answer_relevance":
-            score = llm_answer_relevance(response_generator, question, actual)
-        case "llm_context_precision":
-            score = llm_context_precision(
-                response_generator, question, retrieved_contexts
-            )
-        case "llm_context_recall":
-            score = llm_context_recall(
-                response_generator, question, expected, retrieved_contexts
-            )
-        case _:
-            pass
+        except KeyError:
+            logger.error(f"Unsupported metric type: {metric_type}")
 
     return score
 
@@ -623,21 +113,22 @@ def evaluate_single_prompt(
     map_scores_by_search_type,
     average_precision_for_search_type,
 ):
-    actual = remove_spaces(lower(data.actual))
-    expected = remove_spaces(lower(data.expected))
+    actual = lower_and_strip(data.actual)
+    expected = lower_and_strip(data.expected)
 
     metric_dic = {}
 
     for metric_type in metric_types:
         score = compute_metrics(
-            response_generator,
+            metric_type,
             data.question,
             actual,
             expected,
+            response_generator,
             data.retrieved_contexts,
-            metric_type,
         )
         metric_dic[metric_type] = score
+
     metric_dic["question"] = data.question
     metric_dic["retrieved_contexts"] = data.retrieved_contexts
     metric_dic["actual"] = actual
@@ -810,46 +301,3 @@ def evaluate_prompts(
     )
     draw_hist_df(sum_df, run_id, mlflow_client)
     generate_metrics(config.experiment_name, run_id, mlflow_client)
-
-
-def draw_search_chart(temp_df, run_id, mlflow_client):
-    """
-    Draws a comparison chart of search types across metric types.
-
-    Args:
-        temp_df (pandas.DataFrame): The dataframe containing the data to be plotted.
-        run_id (int): The ID of the current run.
-        mlflow_mlflow_client (mlflow.MlflowClient): The MLflow client to use for logging the chart.
-
-    Returns:
-        None
-    """
-    grouped = temp_df.groupby("search_type")
-    summed_column = grouped.sum().reset_index()
-    fig = sp.make_subplots(rows=len(summed_column.search_type), cols=1)
-    for index, row_data in summed_column.iterrows():
-        search_type = row_data[0]
-        row_data = row_data[1:]
-        df = row_data.reset_index(name="metric_value")
-        df = df.rename(columns={"index": "metric_type"})
-        fig.add_trace(
-            go.Bar(
-                x=df["metric_type"],
-                y=df["metric_value"],
-                name=search_type,
-                offsetgroup=index,
-            ),
-            row=1,
-            col=1,
-        )
-
-        fig.update_xaxes(title_text="Metric type", row=index + 1, col=1)
-        fig.update_yaxes(title_text="score", row=index + 1, col=1)
-    fig.update_layout(
-        font=dict(size=15),
-        title_text="Search type comparison across metric types",
-        height=4000,
-        width=800,
-    )
-    plot_name = "search_type_current_run.html"
-    mlflow_client.log_figure(run_id, fig, plot_name)
